Refactoring: 환승 테이블 첫번째 페이지 로직 분리
